requirements.txt
pip
google-api-python-client
google-auth
pytest
pytest-mock
flake8
alembic
psycopg2-binary
fastapi
sqlalchemy
httpx
fastapi[standard]
requests
python-dotenv
nextflow
awscli
aws
PyJWT
cryptography
boto3
mcp
botocore
# --- Security Patches for Transitive Dependencies ---
filelock>=3.20.3
python-multipart>=0.0.22
starlette>=0.47.2
pyasn1>=0.6.3

# --- Active Agent Developer Kit (ADK) ---
google-adk
google-adk[eval]
google-genai
pytest-asyncio
moto

python-hcl2

asyncpg
greenlet
playwright